+++ src/utils/index.ts
@@ -1,2 +1,19 @@
 export { cn } from './cn.utils';
 export { applyTypoAutoFix } from './typo.utils';
+export { createInkId } from './id.utils';
+export { buildTableHtml } from './table.utils';
+export { getBlockElement, moveBlock, markActiveBlock } from './blocks.utils';
+export {
+  wrapInsertHtml,
+  wrapDeleteHtml,
+  createTrackChange,
+  acceptTrackChangeInHtml,
+  rejectTrackChangeInHtml,
+} from './trackChanges.utils';
+export {
+  createCommentThread,
+  wrapSelectionAsComment,
+  removeCommentMark,
+} from './comments.utils';
+export { InkHistoryStack } from './history.utils';
+export { filterSlashCommands, extractSlashQuery } from './slash.utils';